Mini Project (1) _ Part (1)
This file is related to the first mini project and its first part. In this part, we want to work on the classification of different data in the form of basic algorithms or using Python libraries. In part one, we generate data using the Psychit Learning site and separate different classes from each other using different classifications. In the following, we will classify sample classes with a manual algorithm. We learn how to draw a graph of the loss function and how to determine accuracy in evaluation and training in different ways. Finally, we talk about normalizing data streams and balancing them. Different methods of evaluating test data are also reviewed.